[MOD-243] BBS API
- 팀 게시글 댓글 deleteYn 추가

+++ src/main/java/com/co/kr/modyeo/api/bbs/service/impl/TeamBoardServiceImpl.java
@@ -32,6 +32,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -175,6 +176,14 @@ public class TeamBoardServiceImpl implements TeamBoardService {
                         .errorMessage(BoardErrorCode.NOT_FOUND_REPLY.getMessage())
                         .errorCode(BoardErrorCode.NOT_FOUND_REPLY.getCode())
                         .build());
+
+        if (!Objects.equals(teamReply.getCreatedBy(), memberId)){
+            throw ApiException.builder()
+                    .errorMessage(BoardErrorCode.NOT_DELETE_AUTH.getMessage())
+                    .errorCode(BoardErrorCode.NOT_DELETE_AUTH.getCode())
+                    .status(HttpStatus.FORBIDDEN)
+                    .build();
+        }
 
         teamReply.getTeamArticle().minusReplyCount();
         teamReply.delete();
